chore(es6): transition to es6 - install babel - remove all es5 imports/exports [Finishes #168295003]
