+++ .babelrc
@@ -1,8 +1,5 @@
 {
     "presets": [
       "@babel/preset-env"
-    ],
-    "plugins": [
-      "@babel/plugin-proposal-class-properties"
     ]
 }
