+++ index.js
@@ -1,21 +1,12 @@
-const config = require('config');
-const express  = require('express');
-const auth = require('./routes/authRoutes');
-const mentors = require('./routes/mentorsRoutes');
-const sessions = require('./routes/sessionsRoutes');
-const admin = require('./routes/adminRoutes');
+import express from 'express';
+import auth from './routes/authRoutes';
+import mentors from './routes/mentorsRoutes';
+import sessions from './routes/sessionsRoutes';
+import admin from './routes/adminRoutes';
 
 const app = express();
 
-if(!config.get('privateKey')){
-	console.error('FATAL: No private key defined in config');
-}
 app.use(express.json());
-app.get('/api/v1/auth', (req, res)=>{
-	res.status(200).json({
-		message: 'Welcome to the FreeMentors API'
-	});
-});
 app.use('/api/v1/auth', auth);
 app.use('/api/v1/mentors', mentors);
 app.use('/api/v1/sessions', sessions);
@@ -23,10 +14,6 @@ app.use('/api/v1/user/', admin);
 
 
 const port = process.env.PORT || 3000;
+app.listen(port, ()=> console.log(`Listening on port ${port}...`));
 
-// if(!module.parent){ app.listen(port); }
-if (process.env.NODE_ENV !== 'test') {
-	// app.listen(port);
-	app.listen(port, ()=> console.log(`Listening on port ${port}...`));
-}
-module.exports = app;
+export default app;
