+++ models/mentorsModel.js
@@ -14,4 +14,4 @@ const mentors = [
 	}
 ];
 
-module.exports = mentors;
+export default mentors;
