+++ models/usersModel.js
@@ -40,4 +40,4 @@ const users = [
     }
 ];
 
-module.exports = users;
+export default users;
