+++ routes/adminRoutes.js
@@ -1,8 +1,8 @@
-const adminController = require('../controllers/adminController');
-const userAuth = require('../middleware/userAuth');
-const express = require('express');
-let router = express.Router();
+import adminController from '../controllers/adminController';
+import userAuth from '../middleware/userAuth';
+import express from 'express';
+const router = express.Router();
 
 router.patch('/:userId', userAuth, adminController.upgradeUser);
 
-module.exports = router;
+export default router;
