+++ routes/sessionsRoutes.js
@@ -1,6 +1,6 @@
-const usersController = require('../controllers/usersController');
-const userAuth = require('../middleware/userAuth');
-const express = require('express');
+import usersController from '../controllers/usersController';
+import userAuth from '../middleware/userAuth';
+import express from 'express';
 let router = express.Router();
 
 
@@ -16,4 +16,4 @@ router.post('/:sessionId/review', userAuth, usersController.createSessionReview)
 
 router.delete('/:sessionId/review', userAuth, usersController.adminDeleteReview);
 
-module.exports = router;
+export default router;
